+++ src/LevelInfoLayerMod.hpp
@@ -26,6 +26,16 @@ class $modify(HardcoreLevelInfoLayer, LevelInfoLayer) {
                 auto castedPlayMenu = static_cast<CCMenu*>(playMenu);
                 castedPlayMenu->setOpacity(80);
             }
+            auto leftSideMenu = this->getChildByID("left-side-menu");
+            if (leftSideMenu)
+            {
+                auto copyButton = leftSideMenu->getChildByID("copy-button");
+                if (copyButton)
+                {
+                    auto castedCopyButton = static_cast<CCMenuItemSpriteExtra*>(copyButton);
+                    castedCopyButton->setOpacity(80);
+                }
+            }
         }
         return true;
     }
@@ -40,12 +50,56 @@ class $modify(HardcoreLevelInfoLayer, LevelInfoLayer) {
                 auto castedPlayMenu = static_cast<CCMenu*>(playMenu);
                 castedPlayMenu->setOpacity(80);
             }
+            auto leftSideMenu = this->getChildByID("left-side-menu");
+            if (leftSideMenu)
+            {
+                auto copyButton = leftSideMenu->getChildByID("copy-button");
+                if (copyButton)
+                {
+                    auto castedCopyButton = static_cast<CCMenuItemSpriteExtra*>(copyButton);
+                    castedCopyButton->setOpacity(80);
+                }
+            }
         } else {
             auto playMenu = this->getChildByID("play-menu");
             if (playMenu) {
                 auto castedPlayMenu = static_cast<CCMenu*>(playMenu);
                 castedPlayMenu->setOpacity(255);
             }
+            auto leftSideMenu = this->getChildByID("left-side-menu");
+            if (leftSideMenu)
+            {
+                auto copyButton = leftSideMenu->getChildByID("copy-button");
+                if (copyButton)
+                {
+                    auto castedCopyButton = static_cast<CCMenuItemSpriteExtra*>(copyButton);
+                    castedCopyButton->setOpacity(255);
+                }
+            }
+        }
+    }
+
+    void tryCloneLevel(CCObject* sender) {
+        HardcoreModeManager::getInstance()->currentOnPlaySender = sender;
+        if (m_fields->m_inDifficultyPrompt == false) {
+            int lockoutCheck = HardcoreModeManager::getInstance()->checkIfLockedOutOfLevelId(m_fields->m_currentLevel->m_levelID.value());
+            if (lockoutCheck != -1)
+            {
+                FLAlertLayer::create("Uh oh!", fmt::format("You're locked out of this level for the next {} seconds!", lockoutCheck), "Okay...")->show();
+                return;
+            }
+            geode::createQuickPopup(
+                "GET READY",
+                "CHOOSE YOUR DIFFICULTY",
+                "Normal", "HARDCORE",
+                [](auto, bool btn2) {
+                    HardcoreModeManager::getInstance()->IsInHardcoreMode = btn2;
+                    HardcoreModeManager::getInstance()->currentLevelInfoLayer->onPlay(HardcoreModeManager::getInstance()->currentOnPlaySender);
+                }
+            );
+            m_fields->m_inDifficultyPrompt = true;
+        } else {
+            LevelInfoLayer::tryCloneLevel(sender);
         }
     }
 
@@ -55,7 +109,7 @@ class $modify(HardcoreLevelInfoLayer, LevelInfoLayer) {
             int lockoutCheck = HardcoreModeManager::getInstance()->checkIfLockedOutOfLevelId(m_fields->m_currentLevel->m_levelID.value());
             if (lockoutCheck != -1)
             {
-                FLAlertLayer::create("Uh oh!", fmt::format("You're locked out of this level for thenext {} seconds!", lockoutCheck), "Okay...")->show();
+                FLAlertLayer::create("Uh oh!", fmt::format("You're locked out of this level for the next {} seconds!", lockoutCheck), "Okay...")->show();
                 return;
             }
             //FLAlertLayer* DifficultySelectionAlert = FLAlertLayer::create(newDifficultySelectionAlertLayerProtocol(sender, this), "GET READY", "CHOOSE YOURDIFFICULTY", "Normal", "HARDCORE"); // this    causes a memory leak for the love of god don't use it